fix: split WriteSubnetFile into platform-specific files
